Fix typo in dynamic value syntax in Frame documentation
In Frame.md the dynamic values are surrounded by curly braces - "{parent.w-1}" instead of just "parent.w-1". This lead to the example code not running on any version of Basalt. 

I simply removed the curly braces in this file, not sure if that broke something else or if there are other files that need the same treatment.